finish HouseMemberDocumentService unit tests
