--- src/test/java/com/myhome/services/HouseMemberDocumentServiceTest.java
@@ -1,130 +1,0 @@
-package com.myhome.services;
-
-import com.myhome.domain.HouseMember;
-import com.myhome.domain.HouseMemberDocument;
-import com.myhome.repositories.HouseMemberRepository;
-import com.myhome.services.springdatajpa.HouseMemberDocumentSDJpaService;
-import java.util.Optional;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.MockitoAnnotations;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertNull;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.BDDMockito.given;
-import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.verify;
-
-public class HouseMemberDocumentServiceTest {
-
-  private static final String MEMBER_ID = "test-member-id";
-  private static final String MEMBER_NAME = "test-member-name";
-  private static final HouseMemberDocument MEMBER_DOCUMENT =
-      new HouseMemberDocument("test-file-name", new byte[0]);
-
-  @Mock
-  private HouseMemberRepository houseMemberRepository;
-
-  @InjectMocks
-  private HouseMemberDocumentSDJpaService houseMemberDocumentService;
-
-  @BeforeEach
-  private void init() {
-    MockitoAnnotations.initMocks(this);
-  }
-
-    @BeforeEach
-    private void init() {
-        MockitoAnnotations.initMocks(this);
-    }
-
-  @Test
-  void findMemberDocumentSuccess() {
-    // given
-    HouseMember testMember = new HouseMember(MEMBER_ID, MEMBER_DOCUMENT, MEMBER_NAME, null);
-    given(houseMemberRepository.findByMemberId(MEMBER_ID))
-        .willReturn(Optional.of(testMember));
-    // when
-    Optional<HouseMemberDocument> houseMemberDocument =
-        houseMemberDocumentService.findHouseMemberDocument(MEMBER_ID);
-    // then
-    assertTrue(houseMemberDocument.isPresent());
-    assertEquals(MEMBER_DOCUMENT, houseMemberDocument.get());
-    verify(houseMemberRepository).findByMemberId(MEMBER_ID);
-  }
-
-  @Test
-  void findMemberDocumentNoDocumentPresent() {
-    // given
-    HouseMember testMember = new HouseMember(MEMBER_ID, null, MEMBER_NAME, null);
-    given(houseMemberRepository.findByMemberId(MEMBER_ID))
-        .willReturn(Optional.of(testMember));
-    // when
-    Optional<HouseMemberDocument> houseMemberDocument =
-        houseMemberDocumentService.findHouseMemberDocument(MEMBER_ID);
-    // then
-    assertFalse(houseMemberDocument.isPresent());
-    verify(houseMemberRepository).findByMemberId(MEMBER_ID);
-  }
-
-  @Test
-  void findMemberDocumentMemberNotExists() {
-    // given
-    given(houseMemberRepository.findByMemberId(MEMBER_ID))
-        .willReturn(Optional.empty());
-    // when
-    Optional<HouseMemberDocument> houseMemberDocument =
-        houseMemberDocumentService.findHouseMemberDocument(MEMBER_ID);
-    // then
-    assertFalse(houseMemberDocument.isPresent());
-    verify(houseMemberRepository).findByMemberId(MEMBER_ID);
-  }
-
-  @Test
-  void deleteMemberDocumentSuccess() {
-    // given
-    HouseMember testMember = new HouseMember(MEMBER_ID, MEMBER_DOCUMENT, MEMBER_NAME, null);
-    given(houseMemberRepository.findByMemberId(MEMBER_ID))
-        .willReturn(Optional.of(testMember));
-    // when
-    boolean isDocumentDeleted = houseMemberDocumentService.deleteHouseMemberDocument(MEMBER_ID);
-    // then
-    assertTrue(isDocumentDeleted);
-    assertNull(testMember.getHouseMemberDocument());
-    verify(houseMemberRepository).findByMemberId(MEMBER_ID);
-    verify(houseMemberRepository).save(testMember);
-  }
-
-  @Test
-  void deleteMemberDocumentNoDocumentPresent() {
-    // given
-    HouseMember testMember = new HouseMember(MEMBER_ID, null, MEMBER_NAME, null);
-    given(houseMemberRepository.findByMemberId(MEMBER_ID))
-        .willReturn(Optional.of(testMember));
-    // when
-    boolean isDocumentDeleted = houseMemberDocumentService.deleteHouseMemberDocument(MEMBER_ID);
-    // then
-    assertTrue(isDocumentDeleted);
-    assertNull(testMember.getHouseMemberDocument());
-    verify(houseMemberRepository).findByMemberId(MEMBER_ID);
-    verify(houseMemberRepository, never()).save(testMember);
-  }
-
-  @Test
-  void deleteMemberDocumentMemberNotExists() {
-    // given
-    given(houseMemberRepository.findByMemberId(MEMBER_ID))
-        .willReturn(Optional.empty());
-    // when
-    boolean isDocumentDeleted = houseMemberDocumentService.deleteHouseMemberDocument(MEMBER_ID);
-    // then
-    assertFalse(isDocumentDeleted);
-    verify(houseMemberRepository).findByMemberId(MEMBER_ID);
-    verify(houseMemberRepository, never()).save(any());
-  }
-}
