+++ src/test/java/com/myhome/services/unit/HouseMemberDocumentServiceTest.java
@@ -1,0 +1,285 @@
+package com.myhome.services.unit;
+
+import com.myhome.domain.HouseMember;
+import com.myhome.domain.HouseMemberDocument;
+import com.myhome.repositories.HouseMemberDocumentRepository;
+import com.myhome.repositories.HouseMemberRepository;
+import com.myhome.services.springdatajpa.HouseMemberDocumentSDJpaService;
+
+import java.awt.image.BufferedImage;
+import java.io.ByteArrayOutputStream;
+import java.io.IOException;
+import java.util.Optional;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.MockitoAnnotations;
+import org.springframework.mock.web.MockMultipartFile;
+import org.springframework.test.util.ReflectionTestUtils;
+
+import javax.imageio.ImageIO;
+
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.assertNotEquals;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.BDDMockito.given;
+import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.verify;
+
+public class HouseMemberDocumentServiceTest {
+
+  private static final String MEMBER_ID = "test-member-id";
+  private static final String MEMBER_NAME = "test-member-name";
+  private static final HouseMemberDocument MEMBER_DOCUMENT =
+      new HouseMemberDocument("test-file-name", new byte[0]);
+  private static final int COMPRESSION_BORDER_SIZE_KB = 99;
+  private static final int MAX_FILE_SIZE_KB = 1;
+  private static final long COMPRESSED_IMAGE_QUALITY = (long) 0.99;
+
+  @Mock
+  private HouseMemberRepository houseMemberRepository;
+
+  @Mock
+  private HouseMemberDocumentRepository houseMemberDocumentRepository;
+
+  @InjectMocks
+  private HouseMemberDocumentSDJpaService houseMemberDocumentService;
+
+  @BeforeEach
+  private void init() {
+        MockitoAnnotations.initMocks(this);
+        ReflectionTestUtils.setField(houseMemberDocumentService, "compressionBorderSizeKBytes", COMPRESSION_BORDER_SIZE_KB);
+        ReflectionTestUtils.setField(houseMemberDocumentService, "maxFileSizeKBytes", MAX_FILE_SIZE_KB);
+        ReflectionTestUtils.setField(houseMemberDocumentService, "compressedImageQuality", COMPRESSED_IMAGE_QUALITY);
+    }
+
+  @Test
+  void findMemberDocumentSuccess() {
+    // given
+    HouseMember testMember = new HouseMember(MEMBER_ID, MEMBER_DOCUMENT, MEMBER_NAME, null);
+    given(houseMemberRepository.findByMemberId(MEMBER_ID))
+        .willReturn(Optional.of(testMember));
+    // when
+    Optional<HouseMemberDocument> houseMemberDocument =
+        houseMemberDocumentService.findHouseMemberDocument(MEMBER_ID);
+
+    // then
+    assertTrue(houseMemberDocument.isPresent());
+    assertEquals(MEMBER_DOCUMENT, houseMemberDocument.get());
+    verify(houseMemberRepository).findByMemberId(MEMBER_ID);
+  }
+
+  @Test
+  void findMemberDocumentNoDocumentPresent() {
+    // given
+    HouseMember testMember = new HouseMember(MEMBER_ID, null, MEMBER_NAME, null);
+    given(houseMemberRepository.findByMemberId(MEMBER_ID))
+        .willReturn(Optional.of(testMember));
+    // when
+    Optional<HouseMemberDocument> houseMemberDocument =
+        houseMemberDocumentService.findHouseMemberDocument(MEMBER_ID);
+
+    // then
+    assertFalse(houseMemberDocument.isPresent());
+    verify(houseMemberRepository).findByMemberId(MEMBER_ID);
+  }
+
+  @Test
+  void findMemberDocumentMemberNotExists() {
+    // given
+    given(houseMemberRepository.findByMemberId(MEMBER_ID))
+        .willReturn(Optional.empty());
+    // when
+    Optional<HouseMemberDocument> houseMemberDocument =
+        houseMemberDocumentService.findHouseMemberDocument(MEMBER_ID);
+
+    // then
+    assertFalse(houseMemberDocument.isPresent());
+    verify(houseMemberRepository).findByMemberId(MEMBER_ID);
+  }
+
+  @Test
+  void deleteMemberDocumentSuccess() {
+    // given
+    HouseMember testMember = new HouseMember(MEMBER_ID, MEMBER_DOCUMENT, MEMBER_NAME, null);
+    given(houseMemberRepository.findByMemberId(MEMBER_ID))
+        .willReturn(Optional.of(testMember));
+    // when
+    boolean isDocumentDeleted = houseMemberDocumentService.deleteHouseMemberDocument(MEMBER_ID);
+
+    // then
+    assertTrue(isDocumentDeleted);
+    assertNull(testMember.getHouseMemberDocument());
+    verify(houseMemberRepository).findByMemberId(MEMBER_ID);
+    verify(houseMemberRepository).save(testMember);
+  }
+
+  @Test
+  void deleteMemberDocumentNoDocumentPresent() {
+    // given
+    HouseMember testMember = new HouseMember(MEMBER_ID, null, MEMBER_NAME, null);
+    given(houseMemberRepository.findByMemberId(MEMBER_ID))
+        .willReturn(Optional.of(testMember));
+    // when
+    boolean isDocumentDeleted = houseMemberDocumentService.deleteHouseMemberDocument(MEMBER_ID);
+
+    // then
+    assertTrue(isDocumentDeleted);
+    assertNull(testMember.getHouseMemberDocument());
+    verify(houseMemberRepository).findByMemberId(MEMBER_ID);
+    verify(houseMemberRepository, never()).save(testMember);
+  }
+
+  @Test
+  void deleteMemberDocumentMemberNotExists() {
+    // given
+    given(houseMemberRepository.findByMemberId(MEMBER_ID))
+        .willReturn(Optional.empty());
+    // when
+    boolean isDocumentDeleted = houseMemberDocumentService.deleteHouseMemberDocument(MEMBER_ID);
+
+    // then
+    assertFalse(isDocumentDeleted);
+    verify(houseMemberRepository).findByMemberId(MEMBER_ID);
+    verify(houseMemberRepository, never()).save(any());
+  }
+
+  @Test
+  void updateHouseMemberDocumentSuccess() throws IOException {
+    // given
+    byte[] imageBytes = getImageAsByteArray(10, 10);
+    MockMultipartFile newDocumentFile = new MockMultipartFile("new-test-file-name", imageBytes);
+    HouseMemberDocument savedDocument = new HouseMemberDocument(String.format("member_%s_document.jpg", MEMBER_ID), imageBytes);
+    HouseMember testMember = new HouseMember(MEMBER_ID, MEMBER_DOCUMENT, MEMBER_NAME, null);
+
+    given(houseMemberRepository.findByMemberId(MEMBER_ID))
+            .willReturn(Optional.of(testMember));
+    given(houseMemberDocumentRepository.save(savedDocument))
+            .willReturn(savedDocument);
+    // when
+    Optional<HouseMemberDocument> houseMemberDocument = houseMemberDocumentService.updateHouseMemberDocument(newDocumentFile, MEMBER_ID);
+
+    // then
+    assertTrue(houseMemberDocument.isPresent());
+    assertEquals(testMember.getHouseMemberDocument(), houseMemberDocument.get());
+    verify(houseMemberRepository).findByMemberId(MEMBER_ID);
+    verify(houseMemberDocumentRepository).save(savedDocument);
+    verify(houseMemberRepository).save(testMember);
+  }
+
+  @Test
+  void updateHouseMemberDocumentMemberNotExists() throws IOException {
+    // given
+    byte[] imageBytes = getImageAsByteArray(10, 10);
+    MockMultipartFile newDocumentFile = new MockMultipartFile("new-test-file-name", imageBytes);
+
+    given(houseMemberRepository.findByMemberId(MEMBER_ID))
+            .willReturn(Optional.empty());
+
+    // when
+    Optional<HouseMemberDocument> houseMemberDocument = houseMemberDocumentService.updateHouseMemberDocument(newDocumentFile, MEMBER_ID);
+
+    // then
+    assertFalse(houseMemberDocument.isPresent());
+    verify(houseMemberRepository).findByMemberId(MEMBER_ID);
+    verify(houseMemberDocumentRepository, never()).save(any());
+    verify(houseMemberRepository, never()).save(any());
+
+  }
+
+  @Test
+  void updateHouseMemberDocumentTooLargeFile() throws IOException {
+    // given
+    byte[] imageBytes = getImageAsByteArray(1000, 1000);
+    MockMultipartFile tooLargeDocumentFile = new MockMultipartFile("new-test-file-name", imageBytes);
+    HouseMemberDocument savedDocument = new HouseMemberDocument(String.format("member_%s_document.jpg", MEMBER_ID), imageBytes);
+    HouseMember testMember = new HouseMember(MEMBER_ID, MEMBER_DOCUMENT, MEMBER_NAME, null);
+
+    given(houseMemberRepository.findByMemberId(MEMBER_ID))
+            .willReturn(Optional.of(testMember));
+    given(houseMemberDocumentRepository.save(savedDocument))
+            .willReturn(savedDocument);
+    // when
+    Optional<HouseMemberDocument> houseMemberDocument = houseMemberDocumentService.updateHouseMemberDocument(tooLargeDocumentFile, MEMBER_ID);
+
+    // then
+    assertFalse(houseMemberDocument.isPresent());
+    assertEquals(testMember.getHouseMemberDocument(), MEMBER_DOCUMENT);
+    verify(houseMemberRepository).findByMemberId(MEMBER_ID);
+    verify(houseMemberDocumentRepository, never()).save(any());
+    verify(houseMemberRepository, never()).save(any());
+  }
+
+  @Test
+  void createHouseMemberDocumentSuccess() throws IOException {
+    // given
+    byte[] imageBytes = getImageAsByteArray(10, 10);
+    HouseMemberDocument savedDocument = new HouseMemberDocument(String.format("member_%s_document.jpg", MEMBER_ID), imageBytes);
+    MockMultipartFile newDocumentFile = new MockMultipartFile("new-test-file-name", imageBytes);
+    HouseMember testMember = new HouseMember(MEMBER_ID, MEMBER_DOCUMENT, MEMBER_NAME, null);
+
+    given(houseMemberRepository.findByMemberId(MEMBER_ID))
+            .willReturn(Optional.of(testMember));
+    given(houseMemberDocumentRepository.save(savedDocument))
+            .willReturn(savedDocument);
+    // when
+    Optional<HouseMemberDocument> houseMemberDocument = houseMemberDocumentService.createHouseMemberDocument(newDocumentFile, MEMBER_ID);
+
+    // then
+    assertTrue(houseMemberDocument.isPresent());
+    assertNotEquals(testMember.getHouseMemberDocument().getDocumentFilename(), MEMBER_DOCUMENT.getDocumentFilename());
+    verify(houseMemberRepository).findByMemberId(MEMBER_ID);
+    verify(houseMemberDocumentRepository).save(savedDocument);
+    verify(houseMemberRepository).save(testMember);
+  }
+
+  @Test
+  void createHouseMemberDocumentMemberNotExists() throws IOException {
+    // given
+    byte[] imageBytes = getImageAsByteArray(10, 10);
+    MockMultipartFile newDocumentFile = new MockMultipartFile("new-test-file-name", imageBytes);
+
+    given(houseMemberRepository.findByMemberId(MEMBER_ID))
+            .willReturn(Optional.empty());
+    // when
+    Optional<HouseMemberDocument> houseMemberDocument = houseMemberDocumentService.createHouseMemberDocument(newDocumentFile, MEMBER_ID);
+
+    // then
+    assertFalse(houseMemberDocument.isPresent());
+    verify(houseMemberRepository).findByMemberId(MEMBER_ID);
+    verify(houseMemberDocumentRepository, never()).save(any());
+    verify(houseMemberRepository, never()).save(any());
+  }
+
+  @Test
+  void createHouseMemberDocumentTooLargeFile() throws IOException {
+    // given
+    byte[] imageBytes = getImageAsByteArray(1000, 1000);
+    MockMultipartFile tooLargeDocumentFile = new MockMultipartFile("new-test-file-name", imageBytes);
+    HouseMember testMember = new HouseMember(MEMBER_ID, MEMBER_DOCUMENT, MEMBER_NAME, null);
+
+    given(houseMemberRepository.findByMemberId(MEMBER_ID))
+            .willReturn(Optional.of(testMember));
+    // when
+    Optional<HouseMemberDocument> houseMemberDocument = houseMemberDocumentService.createHouseMemberDocument(tooLargeDocumentFile, MEMBER_ID);
+
+    // then
+    assertFalse(houseMemberDocument.isPresent());
+    assertEquals(testMember.getHouseMemberDocument(), MEMBER_DOCUMENT);
+    verify(houseMemberRepository).findByMemberId(MEMBER_ID);
+    verify(houseMemberDocumentRepository, never()).save(any());
+    verify(houseMemberRepository, never()).save(any());
+  }
+
+  private byte[] getImageAsByteArray(int height, int width) throws IOException {
+    BufferedImage documentImage = new BufferedImage(width, height, BufferedImage.TYPE_INT_RGB);
+    try(ByteArrayOutputStream imageBytesStream = new ByteArrayOutputStream()) {
+      ImageIO.write(documentImage, "jpg", imageBytesStream);
+      return imageBytesStream.toByteArray();
+    }
+  }
+}
